"Changed text of login button from 'Login' to 'Masuk'"
The code diff shows that there is a modification in the index.php file. More specifically, the change is in the text of a link that serves as a button for user login. The text of the button was originally 'Login' and has been changed to 'Masuk'. This is the only change in this code diff.